# docs/reused-openapi/openapi.yaml
openapi: 3.0.3
info:
  title: EV Charging Station API
  description: |
    API για τη διαχείριση σταθμών φόρτισης ηλεκτρικών οχημάτων (EV Charging Stations).
    
    ## Αυθεντικοποίηση
    Η πλειοψηφία των endpoints απαιτεί αυθεντικοποίηση μέσω JWT token.
    Το token λαμβάνεται από το endpoint `/api/v1/auth/signin` και πρέπει να 
    συμπεριλαμβάνεται στο header `Authorization` ως `Bearer <token>`.
    
    ## Ρόλοι Χρηστών
    - **USER**: Βασικός χρήστης - μπορεί να δει φορτιστές, να κάνει κρατήσεις, κ.λπ.
    - **ADMIN**: Διαχειριστής - πλήρης πρόσβαση σε όλες τις λειτουργίες διαχείρισης
  version: 1.0.0
  contact:
    name: softeng25-02 Team
servers:
  - url: http://localhost:8080
    description: ApiGateway development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token που λαμβάνεται από το signin endpoint
  schemas:
...
    Charger:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        connectorType:
          type: string
          enum: [CCS, CHADEMO, TYPE2, TYPE1, SCHUKO]
        maxKW:
          type: number
        status:
          type: string
          enum: [AVAILABLE, IN_USE, OUTAGE]
        pricingProfileId:
          type: integer
          nullable: true
        distanceKm:
          type: number
          description: Απόσταση σε km (μόνο αν δοθεί geo filter)
...
paths:
...
  /api/v1/chargers:
    get:
      tags: [Chargers]
      summary: Αναζήτηση φορτιστών με φίλτρα
      description: |
        Επιστρέφει λίστα φορτιστών με υποστήριξη για:
        - Αναζήτηση κειμένου (όνομα/διεύθυνση)
        - Φιλτράρισμα κατά status, τύπο connector, ισχύ
        - Γεωγραφικό φίλτρο (lat, lng, radius)
        - Ταξινόμηση και σελιδοποίηση
      parameters:
        - name: q
          in: query
          description: Αναζήτηση σε όνομα/διεύθυνση
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [AVAILABLE, IN_USE, OUTAGE]
        - name: connectorType
          in: query
          description: Τύποι connector (comma-separated)
          schema:
            type: string
          example: "CCS,TYPE2"
        - name: minKW
          in: query
          schema:
            type: number
        - name: maxKW
          in: query
          schema:
            type: number
        - name: lat
          in: query
          description: Γεωγραφικό πλάτος (για αναζήτηση σε ακτίνα)
          schema:
            type: number
        - name: lng
          in: query
          description: Γεωγραφικό μήκος
          schema:
            type: number
        - name: radiusKm
          in: query
          description: Ακτίνα αναζήτησης σε km
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
            enum: [distance, power, name]
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Λίστα φορτιστών
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Charger'
                  page:
                    type: integer
                  pageSize:
                    type: integer
                  total:
                    type: integer
  /api/v1/chargers/{id}:
    get:
      tags: [Chargers]
      summary: Λήψη συγκεκριμένου φορτιστή
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Πληροφορίες φορτιστή
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charger'
        '204':
          description: Φορτιστής δεν βρέθηκε
        '400':
          description: Μη έγκυρο ID

  # ==================== POINTS (Alternative API) ====================
...
  /api/v1/reserve/{id}:
    post:
      tags: [Reservations]
      summary: Κράτηση φορτιστή (30 λεπτά default)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID φορτιστή
          schema:
            type: integer
      responses:
        '200':
          description: Κράτηση επιτυχής
          content:
            application/json:
              schema:
                type: object
                properties:
                  pointid:
                    type: string
                  status:
                    type: string
                    example: "reserved"
                  reservationendtime:
                    type: string
                    description: "Format: YYYY-MM-DD HH:MM"
        '400':
          description: Ο χρήστης έχει ήδη ενεργή κράτηση
        '404':
          description: Φορτιστής δεν βρέθηκε
        '409':
          description: Φορτιστής δεν είναι διαθέσιμος
  /api/v1/reserve/{id}/{minutes}:
    post:
      tags: [Reservations]
      summary: Κράτηση φορτιστή για συγκεκριμένο χρόνο
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: minutes
          in: path
          required: true
          description: Διάρκεια κράτησης (max 60 λεπτά)
          schema:
            type: integer
            maximum: 60
      responses:
        '200':
          description: Κράτηση επιτυχής
          content:
            application/json:
              schema:
                type: object
                properties:
                  pointid:
                    type: string
                  status:
                    type: string
                  reservationendtime:
                    type: string
  /api/v1/reserve/{id}/cancel:
    post:
      tags: [Reservations]
      summary: Ακύρωση κράτησης
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID φορτιστή
          schema:
            type: integer
      responses:
        '200':
          description: Κράτηση ακυρώθηκε
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  chargerId:
                    type: integer
        '404':
          description: Δεν βρέθηκε ενεργή κράτηση

  # ==================== SESSIONS ====================
...
  /api/v1/payments/charge:
    post:
      tags: [Payments]
      summary: Χρέωση συνεδρίας φόρτισης
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId]
              properties:
                sessionId:
                  type: integer
                amountEur:
                  type: number
                  description: Ποσό σε EUR (προαιρετικό)
      responses:
        '200':
          description: Χρέωση επιτυχής
        '400':
          description: Μη έγκυρα δεδομένα
...
  /api/v1/admin/chargers:
    post:
      tags: [Admin]
      summary: Δημιουργία νέου φορτιστή
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, providerName, lat, lng, connectorType, maxKW]
              properties:
                name:
                  type: string
                  minLength: 2
                address:
                  type: string
                providerName:
                  type: string
                  minLength: 2
                lat:
                  type: number
                lng:
                  type: number
                connectorType:
                  type: string
                  enum: [CCS, CHADEMO, TYPE2]
                maxKW:
                  type: number
                status:
                  type: string
                  enum: [AVAILABLE, IN_USE, OUTAGE]
      responses:
        '201':
          description: Φορτιστής δημιουργήθηκε
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charger'
        '400':
          description: Μη έγκυρα δεδομένα
  /api/v1/admin/chargers/{id}:
    patch:
      tags: [Admin]
      summary: Ενημέρωση φορτιστή
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
                providerName:
                  type: string
                lat:
                  type: number
                lng:
                  type: number
                connectorType:
                  type: string
                  enum: [CCS, CHADEMO, TYPE2]
                maxKW:
                  type: number
                status:
                  type: string
                  enum: [AVAILABLE, IN_USE, OUTAGE]
      responses:
        '200':
          description: Φορτιστής ενημερώθηκε
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charger'

    delete:
      tags: [Admin]
      summary: Διαγραφή φορτιστή
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Φορτιστής διαγράφηκε
...